add currency repository
