+++ backend/src/currencies/currencies.entity.ts
@@ -1,0 +1,5 @@
+export class Currencies {
+    currency: string;
+    value: number;
+    currentRate: number;
+}
