+++ backend/src/currencies/currencies.repository.ts
@@ -1,0 +1,15 @@
+import { Currencies } from "./currencies.entity";
+
+export class CurrenciesRepository {
+    async getCurrency(currency: string): Promise<Currencies> {
+        return new Currencies();
+    }
+
+    async createCurrency(currency: string): Promise<Currencies> {
+        return new Currencies();
+    }
+
+    async deleteCurrency(currency: string): Promise<void> {
+
+    }
+}
